requirements.txt
Flask
PyYAML
paramiko==3.5.1
Flask-SocketIO
cryptography==42.0.5
PyNaCl==1.5.0
six